Made some dft releated things signed

+++ source/base/math/fourier/dft.cpp
@@ -5,28 +5,28 @@
 
 namespace math {
 
-size_t dft_size(size_t num) {
+int32_t dft_size(int32_t num) {
 	return num / 2 + 1;
 }
 
-void dft_1d(float2* result, const float* source, size_t num) {
+void dft_1d(float2* result, const float* source, int32_t num) {
 	float fn = static_cast<float>(num);
 
 	bool even = true;
 
-	for (size_t k = 0, len = num / 2; k <= len; ++k, even = !even) {
+	for (int32_t k = 0, len = num / 2; k <= len; ++k, even = !even) {
 		float2 sum(0.f);
 
 		float a = -2.f * Pi / fn * static_cast<float>(k);
 
 		if (even) {
-			for (size_t x = 0; x < num; ++x) {
+			for (int32_t x = 0; x < num; ++x) {
 				float b = a * static_cast<float>(x);
 				sum.x += source[x] * std::cos(b);
 		//		sum.y += source[x] * std::sin(b);
 			}
 		} else {
-			for (size_t x = 0; x < num; ++x) {
+			for (int32_t x = 0; x < num; ++x) {
 				float b = a * static_cast<float>(x);
 				sum.x += source[x] * std::cos(b);
 				sum.y += source[x] * std::sin(b);
@@ -41,15 +41,15 @@ void dft_1d(float2* result, const float* source, size_t num) {
 	}
 }
 
-void idft_1d(float* result, const float2* source, size_t num) {
+void idft_1d(float* result, const float2* source, int32_t num) {
 	float fn = static_cast<float>(num);
 
-	for (size_t x = 0; x < num; ++x) {
+	for (int32_t x = 0; x < num; ++x) {
 		float sum = source[0].x;
 
 		float a = -2.f * Pi * static_cast<float>(x) / fn;
 
-		for (size_t k = 1, len = num / 2; k <= len; ++k) {
+		for (int32_t k = 1, len = num / 2; k <= len; ++k) {
 			float b = a * static_cast<float>(k);
 
 			sum += source[k].x * std::cos(b) + source[k].y * std::sin(b);
@@ -61,25 +61,25 @@ void idft_1d(float* result, const float2* source, size_t num) {
 
 // https://www.nayuki.io/page/how-to-implement-the-discrete-fourier-transform
 
-void dft_2d(float2* result, const float* source, size_t width, size_t height) {
-	size_t row_size = dft_size(width);
+void dft_2d(float2* result, const float* source, int32_t width, int32_t height) {
+	int32_t row_size = dft_size(width);
 
 	float2* tmp = new float2[height * row_size];
 
-	for (size_t y = 0; y < height; ++y) {
+	for (int32_t y = 0; y < height; ++y) {
 		dft_1d(tmp + y * row_size, source + y * width, width);
 	}
 
 	float fn = static_cast<float>(height);
 
-	for (size_t x = 0; x < row_size; ++x) {
-		for (size_t k = 0; k < height; ++k) {
+	for (int32_t x = 0; x < row_size; ++x) {
+		for (int32_t k = 0; k < height; ++k) {
 			float2 sum(0.f);
 
 			float a = 2.f * Pi * static_cast<float>(k) / fn;
 
-			for (size_t t = 0; t < height; ++t) {
-				size_t g = t * row_size + x;
+			for (int32_t t = 0; t < height; ++t) {
+				int32_t g = t * row_size + x;
 
 				float angle = a * static_cast<float>(t);
 
@@ -90,7 +90,7 @@ void dft_2d(float2* result, const float* source, size_t width, size_t height) {
 				sum.y += -tmp[g].x * sin_a + tmp[g].y * cos_a;
 			}
 
-			size_t c = k * row_size + x;
+			int32_t c = k * row_size + x;
 			result[c] = sum;
 		}
 	}
@@ -98,8 +98,9 @@ void dft_2d(float2* result, const float* source, size_t width, size_t height) {
 	delete [] tmp;
 }
 
-void dft_2d(float2* result, const float* source, size_t width, size_t height, thread::Pool& pool) {
-	size_t row_size = dft_size(width);
+void dft_2d(float2* result, const float* source, int32_t width, int32_t height,
+			thread::Pool& pool) {
+	int32_t row_size = dft_size(width);
 
 	float2* tmp = new float2[height * row_size];
 
@@ -107,18 +108,18 @@ void dft_2d(float2* result, const float* source, size_t width, size_t height, th
 		for (int32_t y = begin; y < end; ++y) {
 			dft_1d(tmp + y * row_size, source + y * width, width);
 		}
-	}, 0, static_cast<int32_t>(height));
+	}, 0, height);
 
 	pool.run_range([tmp, result, row_size, height](int32_t begin, int32_t end) {
 		float fn = static_cast<float>(height);
 		for (int32_t x = begin; x < end; ++x) {
-			for (size_t k = 0; k < height; ++k) {
+			for (int32_t k = 0; k < height; ++k) {
 				float2 sum(0.f);
 
 				float a = 2.f * Pi * static_cast<float>(k) / fn;
 
-				for (size_t t = 0; t < height; ++t) {
-					size_t g = t * row_size + x;
+				for (int32_t t = 0; t < height; ++t) {
+					int32_t g = t * row_size + x;
 
 					float angle = a * static_cast<float>(t);
 
@@ -129,11 +130,11 @@ void dft_2d(float2* result, const float* source, size_t width, size_t height, th
 					sum.y += -tmp[g].x * sin_a + tmp[g].y * cos_a;
 				}
 
-				size_t c = k * row_size + x;
+				int32_t c = k * row_size + x;
 				result[c] = sum;
 			}
 		}
-	}, 0, static_cast<int32_t>(row_size));
+	}, 0, row_size);
 
 	delete [] tmp;
 }
